# Conflicts:
#	jenetics.ext/build.gradle
#	jenetics.xml/build.gradle
#	jenetics/build.gradle

+++ jenetics.prog/src/test/java/io/jenetics/prog/op/OperationTest.java
@@ -58,7 +58,7 @@ public class OperationTest {
 		Assert.assertEquals(IO.object.fromByteArray(data), object);
 
 		Assert.assertEquals(
-			op.apply(new Double[]{1.0, 2.0}),
+			op.apply(new Double[]{1.0, 2.0}).doubleValue(),
 			3.0
 		);
 	}
